# LSM: Unified manifest