--- a/src/lsm/forest.zig
+++ b/src/lsm/forest.zig
@@ -3,12 +3,15 @@ const builtin = @import("builtin");
 const assert = std.debug.assert;
 const math = std.math;
 const mem = std.mem;
+const log = std.log.scoped(.forest);
 
 const constants = @import("../constants.zig");
 const vsr = @import("../vsr.zig");
 
+const schema = @import("schema.zig");
 const GridType = @import("../vsr/grid.zig").GridType;
 const NodePool = @import("node_pool.zig").NodePool(constants.lsm_manifest_node_size, 16);
+const ManifestLogType = @import("manifest_log.zig").ManifestLogType;
 
 pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
     var groove_fields: []const std.builtin.Type.StructField = &.{};
@@ -71,10 +74,52 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
         }
     }
 
+    const TreeInfo = struct {
+        Tree: type,
+        tree_name: []const u8,
+        tree_id: u16,
+        groove_name: []const u8,
+    };
+
+    const tree_infos: []const TreeInfo = tree_infos: {
+        var tree_infos: []const TreeInfo = &[_]TreeInfo{};
+        for (std.meta.fields(_Grooves)) |groove_field| {
+            const Groove = groove_field.type;
+
+            tree_infos = tree_infos ++ &[_]TreeInfo{.{
+                .Tree = Groove.ObjectTree,
+                .tree_name = groove_field.name,
+                .tree_id = @field(Groove.config.ids, "timestamp"),
+                .groove_name = groove_field.name,
+            }};
+
+            if (Groove.IdTree != void) {
+                tree_infos = tree_infos ++ &[_]TreeInfo{.{
+                    .Tree = Groove.IdTree,
+                    .tree_name = groove_field.name ++ ".id",
+                    .tree_id = @field(Groove.config.ids, "id"),
+                    .groove_name = groove_field.name,
+                }};
+            }
+
+            for (std.meta.fields(Groove.IndexTrees)) |tree_field| {
+                tree_infos = tree_infos ++ &[_]TreeInfo{.{
+                    .Tree = tree_field.type,
+                    .tree_name = groove_field.name ++ "." ++ tree_field.name,
+                    .tree_id = @field(Groove.config.ids, tree_field.name),
+                    .groove_name = groove_field.name,
+                }};
+            }
+        }
+
+        break :tree_infos tree_infos;
+    };
+
     return struct {
         const Forest = @This();
 
         const Grid = GridType(Storage);
+        const ManifestLog = ManifestLogType(Storage);
 
         const Callback = *const fn (*Forest) void;
         const JoinOp = enum {
@@ -90,10 +135,15 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
         join_op: ?JoinOp = null,
         join_pending: usize = 0,
         join_callback: ?Callback = null,
+        callback: ?Callback = null,
 
         grid: *Grid,
         grooves: Grooves,
         node_pool: *NodePool,
+        manifest_log: ManifestLog,
+
+        // TODO JoinOp union?
+        compaction_op: u64 = 0,
 
         pub fn init(
             allocator: mem.Allocator,
@@ -110,6 +160,10 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
             node_pool.* = try NodePool.init(allocator, node_count);
             errdefer node_pool.deinit(allocator);
 
+            var manifest_log =
+                try ManifestLog.init(allocator, grid, .{ .tree_count = tree_infos.len });
+            errdefer manifest_log.deinit(allocator);
+
             var grooves: Grooves = undefined;
             var grooves_initialized: usize = 0;
 
@@ -124,13 +178,7 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
                 const Groove = @TypeOf(groove.*);
                 const groove_options: Groove.Options = @field(grooves_options, groove_field.name);
 
-                groove.* = try Groove.init(
-                    allocator,
-                    node_pool,
-                    grid,
-                    groove_options,
-                );
-
+                groove.* = try Groove.init(allocator, node_pool, grid, groove_options);
                 grooves_initialized += 1;
             }
 
@@ -138,6 +186,7 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
                 .grid = grid,
                 .grooves = grooves,
                 .node_pool = node_pool,
+                .manifest_log = manifest_log,
             };
         }
 
@@ -146,6 +195,7 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
                 @field(forest.grooves, field.name).deinit(allocator);
             }
 
+            forest.manifest_log.deinit(allocator);
             forest.node_pool.deinit(allocator);
             allocator.destroy(forest.node_pool);
         }
@@ -155,6 +205,7 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
                 @field(forest.grooves, field.name).reset();
             }
 
+            forest.manifest_log.reset();
             forest.node_pool.reset();
 
             forest.* = .{
@@ -162,6 +213,7 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
                 .grid = forest.grid,
                 .grooves = forest.grooves,
                 .node_pool = forest.node_pool,
+                .manifest_log = forest.manifest_log,
             };
         }
 
@@ -181,57 +233,146 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
                     return @TypeOf(@field(@as(Grooves, undefined), groove_field_name));
                 }
 
+                pub fn manifest_log_callback(manifest_log: *ManifestLog) void {
+                    const forest = @fieldParentPtr(Forest, "manifest_log", manifest_log);
+                    join(forest);
+                }
+
                 pub fn groove_callback(
                     comptime groove_field_name: []const u8,
                 ) *const fn (*GrooveFor(groove_field_name)) void {
                     return struct {
                         fn groove_cb(groove: *GrooveFor(groove_field_name)) void {
-                            const grooves = @fieldParentPtr(Grooves, groove_field_name, groove);
+                            const grooves: *align(16) Grooves =
+                                @alignCast(@fieldParentPtr(Grooves, groove_field_name, groove));
                             const forest = @fieldParentPtr(Forest, "grooves", grooves);
-
-                            assert(forest.join_op == join_op);
-                            assert(forest.join_callback != null);
-                            assert(forest.join_pending <= std.meta.fields(Grooves).len);
-
-                            forest.join_pending -= 1;
-                            if (forest.join_pending > 0) return;
-
-                            if (join_op == .compacting) {
-                                inline for (std.meta.fields(Grooves)) |field| {
-                                    @field(forest.grooves, field.name).compact_end();
-                                }
-                            }
-
-                            if (join_op == .checkpoint) {
-                                if (Storage == @import("../testing/storage.zig").Storage) {
-                                    // We should have finished all checkpoint writes by now.
-                                    forest.grid.superblock.storage.assert_no_pending_writes(.grid);
-                                }
-                            }
-
-                            const callback = forest.join_callback.?;
-                            forest.join_op = null;
-                            forest.join_callback = null;
-                            callback(forest);
+                            join(forest);
                         }
                     }.groove_cb;
+                }
+
+                fn join(forest: *Forest) void {
+                    assert(forest.join_op == join_op);
+                    assert(forest.join_callback != null);
+                    // +1 for a manifest log compaction.
+                    assert(forest.join_pending <= std.meta.fields(Grooves).len + 1);
+
+                    forest.join_pending -= 1;
+                    if (forest.join_pending > 0) return;
+
+                    if (join_op == .compacting) {
+                        inline for (std.meta.fields(Grooves)) |field| {
+                            @field(forest.grooves, field.name).compact_end();
+                        }
+
+                        const op = forest.compaction_op;
+                        if ((op + 1) % @divExact(constants.lsm_batch_multiple, 2) == 0 and
+                            op >= constants.lsm_batch_multiple and
+                            !forest.grid.superblock.working.vsr_state.op_compacted(op))
+                        {
+                            // Last beat of a half-bar.
+                            forest.manifest_log.forfeit();
+                        }
+                    }
+
+                    if (join_op == .checkpoint) {
+                        if (Storage == @import("../testing/storage.zig").Storage) {
+                            // We should have finished all checkpoint writes by now.
+                            forest.grid.superblock.storage.assert_no_pending_writes(.grid);
+                        }
+                    }
+
+                    const callback = forest.join_callback.?;
+                    forest.join_op = null;
+                    forest.join_callback = null;
+                    callback(forest);
                 }
             };
         }
 
         pub fn open(forest: *Forest, callback: Callback) void {
-            const Join = JoinType(.open);
-            Join.start(forest, callback);
+            assert(forest.callback == null);
+            assert(forest.join_callback == null);
+            forest.callback = callback;
 
             inline for (std.meta.fields(Grooves)) |field| {
-                @field(forest.grooves, field.name).open(Join.groove_callback(field.name));
+                @field(forest.grooves, field.name).open_commence(&forest.manifest_log);
             }
+
+            forest.manifest_log.open(manifest_log_open_event, manifest_log_open_callback);
+        }
+
+        fn manifest_log_open_event(
+            manifest_log: *ManifestLog,
+            level: u7,
+            table: *const schema.ManifestLog.TableInfo,
+        ) void {
+            const forest = @fieldParentPtr(Forest, "manifest_log", manifest_log);
+            assert(forest.callback != null);
+            assert(forest.join_callback == null);
+            assert(level < constants.lsm_levels);
+
+            // TODO Is there a faster (or cleaner) way to map tree-id to the tree?
+            inline for (std.meta.fields(Grooves)) |groove_field| {
+                var groove = &@field(forest.grooves, groove_field.name);
+                const Groove = @TypeOf(groove.*);
+
+                if (groove.objects.config.id == table.tree_id) {
+                    groove.objects.open_table(level, table);
+                    return;
+                }
+
+                if (Groove.IdTree != void) {
+                    if (groove.ids.config.id == table.tree_id) {
+                        groove.ids.open_table(level, table);
+                        return;
+                    }
+                }
+
+                inline for (std.meta.fields(Groove.IndexTrees)) |tree_field| {
+                    const tree = &@field(groove.indexes, tree_field.name);
+                    if (tree.config.id == table.tree_id) {
+                        @field(groove.indexes, tree_field.name).open_table(level, table);
+                        return;
+                    }
+                }
+            }
+
+            log.err("manifest_log_open_event: unknown table in manifest: {}", .{table});
+            @panic("Forest.manifest_log_open_event: unknown table in manifest");
+        }
+
+        fn manifest_log_open_callback(manifest_log: *ManifestLog) void {
+            const forest = @fieldParentPtr(Forest, "manifest_log", manifest_log);
+            assert(forest.callback != null);
+            assert(forest.join_callback == null);
+
+            inline for (std.meta.fields(Grooves)) |field| {
+                @field(forest.grooves, field.name).open_complete();
+            }
+
+            const callback = forest.callback.?;
+            forest.callback = null;
+            callback(forest);
         }
 
         pub fn compact(forest: *Forest, callback: Callback, op: u64) void {
-            // Start a compacting join.
+            assert(forest.callback == null);
+
             const Join = JoinType(.compacting);
             Join.start(forest, callback);
+
+            forest.compaction_op = op;
+
+            if (op % @divExact(constants.lsm_batch_multiple, 2) == 0 and
+                op >= constants.lsm_batch_multiple and
+                !forest.grid.superblock.working.vsr_state.op_compacted(op))
+            {
+                // This is the first beat of a bar that we have not compacted already.
+                forest.join_pending += 1;
+                forest.manifest_log.reserve();
+                forest.manifest_log.compact(Join.manifest_log_callback);
+            }
 
             inline for (std.meta.fields(Grooves)) |field| {
                 @field(forest.grooves, field.name).compact(Join.groove_callback(field.name), op);
@@ -239,17 +380,27 @@ pub fn ForestType(comptime Storage: type, comptime groove_cfg: anytype) type {
         }
 
         pub fn checkpoint(forest: *Forest, callback: Callback) void {
+            assert(forest.callback == null);
+            forest.callback = callback;
+
             if (Storage == @import("../testing/storage.zig").Storage) {
                 // We should have finished all pending io before checkpointing.
                 forest.grid.superblock.storage.assert_no_pending_writes(.grid);
             }
 
-            const Join = JoinType(.checkpoint);
-            Join.start(forest, callback);
-
             inline for (std.meta.fields(Grooves)) |field| {
-                @field(forest.grooves, field.name).checkpoint(Join.groove_callback(field.name));
+                @field(forest.grooves, field.name).assert_between_bars();
             }
+
+            forest.manifest_log.checkpoint(manifest_log_callback);
+        }
+
+        fn manifest_log_callback(manifest_log: *ManifestLog) void {
+            const forest = @fieldParentPtr(Forest, "manifest_log", manifest_log);
+            const callback = forest.callback.?;
+
+            forest.callback = null;
+            callback(forest);
         }
     };
 }