## Summary

- Store tree id in TableInfo in the manifest log blocks on disk.
- Store `TableInfo` of every tree in each manifest log block.
- Move the `ManifestLog` from one-per-`Tree` to one-per-`Forest`.

#### Results

- Average case manifest log writes for a checkpoint: 21 blocks to 2 or 3 blocks (ballpark).
- Worst-case manifest log writes for a compaction half-bar: 63 blocks to 6 blocks.
- Worst-case manifest log writes for a checkpoint: 63 blocks to 6 blocks.
- Memory owned by manifest log: 63 blocks to 6 blocks.
- `state_machine.checkpoint()` latency drops by ~70% in the benchmark (on my laptop). (Note that this is an _extremely_ small sample size, though -- the benchmark only checkpoints once per run.)

<details>
<summary>(Checkpoint latency samples)</summary>

(Measuring `state_machine.checkpoint()` from invocation to callback.)

```
main:
0.041102385s
0.044551919s
0.033833146s
0.033007621s

unified manifest:
0.010172937s
0.012081376s
0.009965735s
```
</details>

## Context

Each tree's `TableInfo` is currently (abbreviated):

    fn TableInfoType(Key) type
        return extern struct {
            checksum: u128,
            address: u64,
            flags: u64, // (Currently unused.)
            snapshot_min: u64,
            snapshot_max: u64,
            key_min: Key,
            key_max: Key,
        };
    }

Some relevant constants:

    block_size                             = 64KiB
    lsm_batch_multiple                     = 64 ops
    journal_slot_count                     = 1024 ops
    storage_size_max                       = 16TiB
    tree_count                             = 21 trees
    lsm_levels                             = 7
    bars_per_checkpoint                    = 16 = 1024 / 64

    // Worst-case: fully-loaded LSM, with every possible compaction running every bar:
    compactions_pre_bar_per_tree           = 112  = lsm_levels * bars_per_checkpoint
    compactions_per_bar                    = 2352 = lsm_levels * bars_per_checkpoint * tree_count

    // This is a pretty arbitary constant – we can tune it up or down (preferrable down) as necessary.
    // But it is part of the disk layout, so we need to pick a value and stick with it.
    // (See superblock_trailer_manifest_size_max.)
    maximum superblock manifest size       = 36_184_064B = ~34.5MiB

And currently:

    SuperBlock.Manifest.BlockReferenceSize = 40B = @sizeof(tree_id) + @sizeof(checksum) + @sizeof(address)

    // Pretending that every manifest log block is full (which they aren't in practice – see below):
    maximum manifest log blocks            = ⌊36_184_064B / 40B⌋ = 904_601 blocks
    maximum manifest log bytes             = 904_601 * block_size = ~55.2GiB
    // Pretending that all Keys are 16B (as they are for the majority of trees).
    maximum manifest updates               = 723_680_800 updates = 800 updates/block * 904_601 blocks

    Trees:
    | @sizeof(Tree.Key) | @sizeof(TableInfo) | manifest updates/block | Tree Count |
    |                8B |                64B |                    992 |          4 |
    |               16B |                80B |                    800 |         12 |
    |               32B |               112B |                    576 |          5 |

However, the _updates-per-block_ numbers (and thus _maximum manifest updates_) above are a misleading ballpark:

- Different trees have different manifest log blocks.
- Different levels of the same tree have the same manifest log blocks.
- Lower bound: the LSM writes one manifest log block per tree (that is written to during the checkpoint) per checkpoint. (21 trees → 21 blocks written/checkpoint).
- Not counting manifest log compaction, each of those blocks has _at most_ 112 (`compactions_pre_bar_per_tree`) blocks – a write/space amplification of ~7 (`TableInfo/block ÷ compactions_pre_bar_per_tree`).
- Usually we don't compact every tree of every level. So the blocks are even emptier! (More space/write amplification).
- The amplification isn't a huge deal, but those 21 writes are a lower bound, and they occur during the checkpoint, contributing to the checkpoint latency spike.
- Additionally, the incomplete manifest log blocks increase the amount of manifest log compaction work outside of checkpoints.

## Unified Manifest Log

Instead of dividing the manifest log between trees, we could put all updates to all trees in the same manifest log blocks.

For this to work, `TableInfo` would need:

- To include a tree identifier (recently migrated from `u128` to `u16`).
- To accommodate any table, the `Key`s must be padded to the largest key size of any tree (`32B`).
- Then `@sizeOf(TableInfo) == 112B` _for every tree_.
- Lower bound: the LSM writes one manifest log block total per checkpoint.

On the superblock manifest side:
- No need to store the `tree_id` in the superblock manifest anymore, so `SuperBlock.Manifest.BlockReferenceSize == 24B` (down from `40B`).

Then:

    // Using the same value as a base, then working backwards:
    maximum manifest updates               = 723_680_800 updates
    maximum manifest log blocks            = 722_680_800 updates ÷ 576 updates/block = 1_256_390 blocks

    // 24B = @sizeof(checksum) + @sizeof(address)
    SuperBlock.Manifest.BlockReferenceSize = 24B
    maximum superblock manifest size       = 1_256_390 blocks * 24 bytes/block = 30_153_360B = 28.8MiB

This means that despite paying for the largest key size for trees with small keys, the overall superblock manifest size shrinks as we drop the tree id.
(Note that this is because we currently encode tree id as a `u128` even though it fits in a `u16`.)

#### Pros

- Smaller superblock manifest upper bound.
- Smaller superblock manifest in practice (due to less space amplification).
  - This means less data for checkpoints to write.
  - This means a little less data for state sync to move.
- Lower manifest log space/read/write amplification.
  - This reduced the amount of work necessary when starting the replica (i.e. opening the LSM).
- Fewer manifest log writes, especially at checkpoint.
- Faster recovery (fewer, denser manifest blocks to read).
- Lower `ManifestLog` RAM usage (approximately `21*3*64KiB ≈ 4MiB` → `6*64KiB = 384KiB`).
- De-specialize `ManifestLog`.

#### Cons

- We can't add a tree with a key larger than `32B`.
  - (Well, we can, but we would need to re-version the manifest block format).
  - (For secondary indexes, that means the maximum is `8B timestamp + 24B indexed key`).
  - (If we anticipate one a larger key, we could pad in advance, but the above calculations all assume 32B is the limit.)
- Unified manifest log compaction can't separate hot and cold manifests – all of the entries in a manifest log block are rewritten when it is compacted.
  - (So if one tree is modified infrequently, while the other is modified often, rewriting the cold tree's manifest entries is now write amplification).

---

<details>
<summary>Unified Manifest Level?</summary>

## Unified Manifest Level?

All of the above is entirely about the manifest _on disk_.

For completeness: **What about the in-memory manifest?**

Two options:
1. Specialized. (This is what we already have implemented.)
2. Unified.

I lean towards _specialized manifest levels_.
But either way, the decision doesn't impact the storage format, so it is flexible in the future.

### Specialized

Right now we have per-tree `TableInfo`, with a per-tree specialized `ManifestLevel`.

#### Pros
- This is more RAM-efficient – no overhead due to padding keys up to 32B.
- (Also, this is already implemented.)

#### Cons
- We need separate `TableInfo` type(s) from the `TableInfo` that is serialized into the manifest log blocks. (`schema.TableInfo` vs `ManifestLevel(Tree).TableInfo`)
- Per-tree specialization of `ManifestLevel` (i.e. more code generation).

### Unified

Change `ManifestLevel` (and `SegmentedArray`) to take a `*const fn` for `compare_keys` (instead of a `comptime fn callconv(.Inline)`).

#### Pros
- No pre-tree specialization of `ManifestLevel` (i.e. less code generation).

#### Cons
- More RAM overhead – tables with 8B or 16B keys must still pad their keys to 32B.
- `compare_keys()` can't be inlined.
- My guess is that despecializing `SegmentedArray` will have a noticeable (CPU) performance impact when searching levels with many tables.

</details>