@matklad 2023-08-30
> We can't add a tree with a key larger than 32B

Here, could we in theory make manifest log entries variable in size? I think we can --- the data structure here is a log, it is processed sequentially and doesn't need random access. If we can make entries var-sized, then yeah, I don't think are any forward compatibility concerns here --- if at some point we decide to change `TableInfo` for a particular table size, we will be able to do that.