@sentientwaffle 2023-08-30
> Here, could we in theory make manifest log entries variable in size? I think we can --- the data structure here is a log, it is processed sequentially and doesn't need random access. If we can make entries var-sized, then yeah, I don't think are any forward compatibility concerns here --- if at some point we decide to change `TableInfo` for a particular table size, we will be able to do that.

:thinking: I think I see what you mean. As an added bonus, only 5/21 of our trees use the maximum key size -- so in practice we could pack more than 576 (variable-sized) `TableInfo`s into each block. (We still have to allocate enough blocks to handle the worst (maximum-key-size) case.)

But: It would make APIs more complicated, since we can't have a `schema.ManifestLog.TableInfo` anymore -- we would need to pass + encode the tree's specialized `TableInfo`s.

Also the manifest log currently schedules via `entry_count < entry_count_max`. But with variable-sized `TableInfo`s, `entry_count_max` isn't that useful -- we would rather just check whether the next `TableInfo` can fit into the current block.

And when computing `blocks_count_max` (the maximum number of blocks that the manifest log must ever buffer) we need to assume the worst case, where every `TableInfo` has the largest key size.

Overall I'm not sure whether this is worth it right now, since we don't anticipate requiring larger keys.