@matklad 2023-08-30
Yeah, we definitely don't need this right now. I am thinking more from an angle of "would we regret unified manifest if, at some point, we would like to have table-specific log entries of different sizes". Seems the answer is now, because you could have _both_ unified manifest, and also table-specific log entries.